+++ src/intefaces/Client/ClientInterface.ts
@@ -1,0 +1,9 @@
+import type { ClientFormInteface } from './ClientFormInteface'
+
+export interface ClientInterface extends ClientFormInteface {
+  id: number
+  user: {
+    id: number
+  }
+  name: string
+}
