+++ src/intefaces/Client/PageFormClientInteface.ts
@@ -1,0 +1,5 @@
+export default interface PageFormClientInteface {
+  send: () => void
+  titlePage: string
+  blockTitle: string
+}
